requirements.txt
# FastAPI and server dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2

# Testing
pytest==8.3.3
httpx==0.27.2
pytest-asyncio

# LlamaIndex core
llama-index-core

# LLM provider - Gemini
llama-index-llms-gemini
llama-index-llms-google-genai
google-generativeai

# Embeddings - HuggingFace
llama-index-embeddings-huggingface

# Vector store - Qdrant
llama-index-vector-stores-qdrant
qdrant-client

# Utilities
python-dotenv
pyyaml
